session2_3/pyproject.toml
[project]
name = "session2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
]
